Handle key events for songpanel in songpanel instead of parent form.
Make sure songpanel gets focus when user clicks in it or other panels/screens are closed.
This removes suppressing of keys when other controls have focus so you can type freely in the web browser address bar for example.
But leave key suppressing on so that the track/song props toggle buttons won't be clicked when s/t/space is pressed. Alt-s/t still works.