Fix bug which caused dragging destination of TrackList items to not be highlighted
If the TrackList control is created small and then enlarged by enlarging/maximizing the application window, GDI drawing won't work outside the initial bounds. A quick solution was to set the initial window height large enough. The height can't be set larger than the current screen resolution, but as a workaround an invisible label control is placed at y = 4360 (to handle 8k screens). This will automatically resize the window and all its docked content including the TrackList.
Also, remove the old "debugLabel" which is not used anymore.